KB-1270 Refactor GrantButton
