+++ src/app/[locale]/(private)/module/studbonuspointsrights/components/grant-button.tsx
@@ -1,43 +1,17 @@
-'use client';
-
-import { useState } from 'react';
 import { useTranslations } from 'next-intl';
 import { Plus } from 'lucide-react';
 import { Button } from '@/components/ui/button';
-import {
-  SbpLoadCatalogItem,
-  SbpStudyYearCatalogItem,
-  SbpSubdivisionCatalogItem,
-} from '@/types/models/sbp-rights';
-import { GrantRightsDialog } from './grant-rights-dialog';
+import { Link } from '@/i18n/routing';
 
-interface Props {
-  loads: SbpLoadCatalogItem[];
-  subdivisions: SbpSubdivisionCatalogItem[];
-  years: SbpStudyYearCatalogItem[];
-}
-
-/**
- * Header button that opens the Grant Rights dialog. Holds the open state
- * locally so the page (a server component) doesn't need to.
- */
-export function GrantButton({ loads, subdivisions, years }: Props) {
+export function GrantButton() {
   const t = useTranslations('private.studbonuspointsrights.grant');
-  const [open, setOpen] = useState(false);
 
   return (
-    <>
-      <Button onClick={() => setOpen(true)} className="flex items-center gap-2">
-        <Plus className="h-4 w-4" />
+    <Button asChild size="small">
+      <Link href="/module/studbonuspointsrights/grant">
+        <Plus />
         {t('button')}
-      </Button>
-      <GrantRightsDialog
-        open={open}
-        onOpenChange={setOpen}
-        loads={loads}
-        subdivisions={subdivisions}
-        years={years}
-      />
-    </>
+      </Link>
+    </Button>
   );
 }
