+++ src/messages/uk.json
@@ -536,6 +536,7 @@
         "userSearchPlaceholder": "Почніть вводити ПІБ",
         "userMinChars": "Введіть щонайменше 2 символи.",
         "userNotFound": "Користувача не знайдено.",
+        "userSearching": "Пошук...",
         "scope": "Рівень",
         "scopeFaculty": "Факультет",
         "scopeUniversity": "Університет",
